requirements.txt
pyserial
chess